Massive rework of the entire application
Cleaned up the data model a lot. Made id's be UUIDs. Removed track and
show from the database because channels and episodes make more sense for
it being a podcast.

One idea that has sprang up from this is the ability to create custom
playlists from other's shows. We could make shows be optionally linked
to channels, but for now. This will do.